+++ src/components/utils.ts
@@ -660,7 +660,7 @@ export function deprecationAssert(condition: boolean, oldProperty: string, newPr
  * @param propertyKey - method or accessor name
  * @param descriptor - property descriptor
  */
-export function cacheable<Target, Value, Arguments extends any[] = any[]>(
+export function cacheable<Target, Value, Arguments extends unknown[] = unknown[]>(
   target: Target,
   propertyKey: string,
   descriptor: PropertyDescriptor
@@ -672,7 +672,7 @@ export function cacheable<Target, Value, Arguments extends any[] = any[]>(
   /**
    * Override get or value descriptor property to cache return value
    *
-   * @param args
+   * @param args - method args
    */
   descriptor[propertyToOverride] = function (...args: Arguments): Value {
     /**
@@ -688,12 +688,12 @@ export function cacheable<Target, Value, Arguments extends any[] = any[]>(
   /**
    * If get accessor has been overridden, we need to override set accessor to clear cache
    *
-   * @param value
+   * @param value - value to set
    */
   if (propertyToOverride === 'get' && descriptor.set) {
     const originalSet = descriptor.set;
 
-    descriptor.set = function (value: any): void {
+    descriptor.set = function (value: unknown): void {
       delete target[cacheKey];
 
       originalSet.apply(this, value);